requirements.txt
anyio==4.6.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
et-xmlfile==1.1.0
Flask==3.0.3
googlemaps==4.10.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
lxml==5.3.0
MarkupSafe==2.1.5
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parsel==1.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tzdata==2024.2
urllib3==2.2.3
w3lib==2.2.1
Werkzeug==3.0.4
